### 2025-02-18 #687 @bhroston
fcmconfr: A Package for Analysis of Fuzzy Cognitive Maps (FCMs)
Submitting Author Name: Name
Submitting Author Github Handle: @bhroston
Repository:  https://github.com/bhroston/fcmconfr
Submission type: Pre-submission

---

```
Package: fcmconfr
Title: Fuzzy Cognitive Map Analysis in R
Version: 0.2.3
  person(
    "Ben", "Roston",
    email = "bhroston@vt.edu",
    comment = c(ORCID = "0000-0001-9201-7579")
  ),
  person(
    "Meg", "Rippy",
    email = "mrippy@vt.edu",
    role = c("aut", "cph", "ths"),
    comment = c(ORCID = "0000-0002-0575-8342")
  ))
Description: fcmconfr streamlines the process of conducting dynamic simulations
  using Fuzzy Cognitive Maps (FCMs). The package supports multiple FCM types
  (conventional, Interval-Value Fuzzy Number (IVFN) and Triangular Fuzzy Number
  (TFN)) and provides tools for model aggregation and the propagation of
  uncertainty from individual FCMs through to the aggregate.
License: GPL (>= 3) + file LICENSE
    covr (>= 3.6.4),
    knitr (>= 1.44),
    rmarkdown (>= 2.25),
    vdiffr (>= 1.0.8)
RdMacros: Rdpack
    Rdpack (>= 2.6.1),
    rlang (>= 1.1.4),
    cli (>= 3.6.3),
    methods (>= 4.3.2),
    data.table (>= 1.15.4),
    stats (>= 4.3.2),
    utils (>= 4.3.2),
    parallel (>= 4.3.2),
    foreach (>= 1.5.2),
    doSNOW (>= 1.0.20),
    pbapply (>= 1.7.2),
    Matrix (>= 1.6.1.1),
    tidyr (>= 1.3.0),
    shinyWidgets (>= 0.8.7),
    bslib (>= 0.8.0),
    graphics (>= 4.3.2),
    tibble (>= 3.2.1),
    ggplot2 (>= 3.4.2),
    visNetwork (>= 2.1.2),
    igraph (>= 1.5.1),
    magrittr (>= 2.0.3),
    shiny (>= 1.7.5)

```

Exploratory Data Analysis (EDA) and Summary Statistics: This package is intended to give users tools to perform an array of FCM analyses (simulation, aggregation, and error propagation) and provides visualization functions and various GUIs to interact with the data. EDA is broad enough that I could make an argument that this package fits within its bounds, but I don't necessarily think of fcmconfr as primarily an EDA package.
Bayesian and Monte Carlo Routines: This package uses Monte Carlo routines to estimate uncertainty in its analysis, though I would not say that is its sole primary aim.
I would greatly appreciate advice about whether this package is within the scope of these or any of the above statistical package categories.

Not yet, but will complete once the correct category(ies) is/are determined.

This package targets researchers interested in analyzing Fuzzy Cognitive Maps in their projects. FCM research has a significant barrier-to-entry because (a) it can be challenging to translate the math into code, (b) it can be even harder to incorporate uncertainty calculations into FCM analyses, and (c) there is not a consensus in the literature about one "best" way to analyze FCMs. This package is designed to lower the barrier-to-entry to FCM analysis while at the same time offering state-of-the-literature functionality for advanced users.

fcmconfr shares some functionality with the existing fcm package, but it is broader in functionality. The fcm package provides a straightforward function to simulate FCMs. fcmconfr also performs FCM simulation, and also provides FCM aggregation functionality alongside tools to incorporate uncertainty into its analysis.

Probably N/A?

The pkgcheck Github action says this package has a lot of imports and gives the curious eyes emoji saying it might be worth looking into. I agree there are a lot, but many seem like core packages like rlang, stats, graphics, etc. Should I address this note?
