### 2025-02-18 #687 @mpadge
@bhroston Thanks for your pre-submission inquiry. You package indeed seems like a good candidate for statistical peer review. Our general rule there is that a category applies if you think you can comply with at least 50% of all standards. I had a look through your package alongside the Bayesian and Monte Carlo standards, and it does indeed seem like that would be the case. Obviously the _Bayesian_ bit of those standards wouldn't apply, which will make a few standards not applicable, but the majority seem like they'd still apply. It also seems that your package should also be considered an EDA package.

I suggest that you initially grab lists of all of the Bayesian and Monte Carlo as well as EDA standards using [the `srr` package as described here](https://stats-devguide.ropensci.org/pkgdev.html#pkgdev-srr). Use those as initial checks list to confirm that you'll be able to comply with at least half of each. Let us know if so, and then proceed both to comply with all applicable standards, and to document compliance within your package using the [`srr` system](https://docs.ropensci.org/srr/articles/srr-stats.html).

---

In regard to your question about numbers of imports: Yes indeed, 22 imports is a lot. One way to reduce that number is to move a selection from _Imports_ to _Suggests_. Packages associated with visualisation routines, such as `visNetwork` and `shinyWidgets`, and maybe even `igraph`, could easily be _Suggests_. Each routine which uses those should then start with `importNamespace("visNetwork")` (or whichever) to load only where required. More detail in https://r-pkgs.org/dependencies-mindset-background.html#sec-dependencies-imports-vs-suggests. (And we only just implemented that check this month, great to see it already in action :+1:)

Thanks again for starting the process here!
