### 2025-03-25 #687 @mpadge
@bhroston That's great news. Don't worry too much about `autotest`. I've just updated the [*Stats DevGuide* chapter](https://stats-devguide.ropensci.org/pkgdev.html#pkgdev-autotest) to include this statement:

> Note that use of [`autotest`](https://github.com/ropensci-review-tools/autotest) is recommended yet not strictly required. The package is intended to aid package development and preparation for submission, yet it is still experimental, and it may not be possible for all packages to generate clean (`NULL`) results.

We do no checking of `autotest` results on submission, only `srr` and `pkgcheck`. As long as they both confirm that everything is okay, then you should be good. You can also call `@ropensci-review-bot check package` here to confirm. If good, then I'd suggest asking current Editor-in-Chief @maurolepore whether he thinks it's okay to proceed.
