### 2025-03-31 #687 @bhroston
The review bot's package checks returned successful, so we can confirm the package is ready to be submitted in that sense.

@maurolepore This package successfully passes pkgcheck as confirmed by the review bot above. I've been discussing with @mpadge about whether it is ready for submission and they recommended I reach out for your input.

A note on the state of the package: The code is complete, but there may be minor updates to documentation and the primary vignette to ensure all parameters and algorithms are described clearly and accurately. Also, I am a little over a month away from my PhD thesis and one chapter will be a JOSS-style paper describing the package that may be included as a vignette later on.

Should I submit the package for review now or wait until my advisor and I have finalized the vignettes and supplementary paper?

Thanks,
Ben
