### 2025-12-16 #687 @bhroston
@mpadge Thanks for your feedback along the way! I'd be happy to wait until after the New Year to really get things going. Just for clarification, should I wait to open a full submission now or wait until Jeff is back in January?
