### 2026-01-05 #687 @jhollist
@bhroston Following up with @mpadge comment.

I would like to invite you to open a full submission to rOpenSci for your package, `fcmconfr`.

Thank you for your interest in rOpenSci.  Look forward to continuing the review process.
